+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added sequence sort option and new default sort functionality to search (DR-3609)
 - Connected collection landing page (metadata) to API (DR-3362)
 - Added collection structure component (DR-3545)
+- Added deprecated division slug redirects
 
 ## [0.4.1] 2025-05-16